# Conflicts:
#	src/main/java/dao/PedidoDao.java
#	src/main/java/daoImp/ComisionDaoImp.java
#	src/main/java/daoImp/ComisionFijaDaoImp.java
#	src/main/java/daoImp/ComisionPorcentualDaoImp.java
#	src/main/java/daoImp/FormaPagoDaoImp.java
#	web/index.html

+++ src/main/java/dao/ComisionFijaDao.java
@@ -2,5 +2,5 @@ package dao;
 
 import model.ComisionFija;
 
-public interface ComisionFijaDao extends GenericDao<ComisionFija,Integer> {
+public interface ComisionFijaDao extends GenericDao<ComisionFija,Long> {
 }
