+++ src/main/java/dao/ProductoDao.java
@@ -2,5 +2,5 @@ package dao;
 
 import model.Producto;
 
-public interface ProductoDao extends GenericDao<Producto, Integer> {
+public interface ProductoDao extends GenericDao<Producto, Long> {
 }
