+++ src/main/java/dao/TarjetaDao.java
@@ -3,5 +3,5 @@ package dao;
 import model.Tarjeta;
 
 
-public interface TarjetaDao  extends GenericDao<Tarjeta, Integer> {
+public interface TarjetaDao  extends GenericDao<Tarjeta, Long> {
 }
